added now ts, nonce, signature to delete

+++ anontion/src/main/account/java/com/anontion/account/controller/AccountDeleteController.java
@@ -2,10 +2,13 @@ package com.anontion.account.controller;
 
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.Optional;
 import java.util.UUID;
+import java.util.concurrent.TimeUnit;
 
+import org.bouncycastle.crypto.params.ECPublicKeyParameters;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
@@ -19,15 +22,20 @@ import com.anontion.asterisk.repository.AsteriskAuthRepository;
 import com.anontion.asterisk.repository.AsteriskEndpointRepository;
 import com.anontion.common.dto.response.ResponseDTO;
 import com.anontion.common.dto.response.Responses;
+import com.anontion.common.misc.AnontionConfig;
 import com.anontion.common.misc.AnontionLog;
 import com.anontion.common.misc.AnontionStrings;
+import com.anontion.common.misc.AnontionTime;
 import com.anontion.common.security.AnontionSecurity;
+import com.anontion.common.security.AnontionSecurityECDSA;
 import com.anontion.models.account.model.AnontionAccount;
 import com.anontion.models.account.repository.AnontionAccountRepository;
 import com.anontion.models.application.model.AnontionApplication;
 import com.anontion.models.application.model.AnontionApplicationId;
 import com.anontion.models.application.repository.AnontionApplicationRepository;
 import com.anontion.services.service.AccountService;
+import com.github.benmanes.caffeine.cache.Cache;
+import com.github.benmanes.caffeine.cache.Caffeine;
 
 @RestController
 public class AccountDeleteController {
@@ -50,9 +58,39 @@ public class AccountDeleteController {
   @Autowired
   private AccountService accountService;
  
+  private Cache<String, Boolean> nonceCache = Caffeine.newBuilder()
+      .expireAfterWrite(AnontionConfig._REQUEST_TS_VALIDITY_MARGIN, TimeUnit.SECONDS)
+      .build();
+  
   @DeleteMapping("/account/{id}")
-  public ResponseEntity<ResponseDTO> deleteAccount(@PathVariable("id") String id) {
+  public ResponseEntity<ResponseDTO> deleteAccount(@PathVariable("id") String id, @RequestParam("nowTs") Long nowTs, @RequestParam("nonce") String nonce, @RequestParam("signature") String signature) {
 
+    if (nowTs <= 0 ||
+        nonce.isBlank() ||
+        signature.isBlank()) {
+      
+      return Responses.getBAD_REQUEST(
+          "Parameters failed validity check.", 
+          "Transaction rejected.");
+    }
+    
+    Long diff = nowTs - AnontionTime.tsN();
+
+    if (diff > AnontionConfig._REQUEST_TS_VALIDITY_MAX ||
+        diff < AnontionConfig._REQUEST_TS_VALIDITY_MIN) {
+
+      return Responses.getBAD_REQUEST(
+          "Invalid parameters time", 
+          "bad nowTs"); 
+    }
+    
+    if (nonce.isBlank() || nonceCache.getIfPresent(nonce) != null) {
+
+      return Responses.getBAD_REQUEST("Bad message");
+    }
+    
+    nonceCache.put(nonce, true);
+    
     String[] tokens = AnontionStrings.split(id, ":");
     
     if (tokens.length < 3) {
@@ -144,15 +182,41 @@ public class AccountDeleteController {
       }
     }
     
-    if (!accountService.deleteTxApplicationAndAccountAndEndpoint(
-        applicationO.orElse(null), 
-        accountO.orElse(null),
-        (endpoint0 == null ? null : endpoint0.orElse(null)),
-        auths,
-        aors)) {
+    if (!accountO.isEmpty()) {
+
+      AnontionAccount account = accountO.get();
+
+      String text = nowTs.toString() + nonce;
+
+      ECPublicKeyParameters publicKey = 
+          AnontionSecurityECDSA.decodeECPublicKeyParametersFromBase64XY(account.getClientPub());
+
+      if (publicKey == null) {
+
+        return Responses.getBAD_REQUEST(
+            "Bad pub.", 
+            "Pub invalid.");
+      }
+
+      if (!AnontionSecurityECDSA.checkSignature(text, signature, publicKey)) {
+
+        return Responses.getBAD_REQUEST(
+            "Bad signature.", 
+            "Signature invalid.");
+      } 
+
+      if (!accountService.deleteTxApplicationAndAccountAndEndpoint(
+          applicationO.orElse(null), 
+          accountO.orElse(null),
+          (endpoint0 == null ? null : endpoint0.orElse(null)),
+          auths,
+          aors)) {
+
+        return Responses.getBAD_REQUEST(
+            "Failed application, account and/or asterisk update.");
+      }
       
-      return Responses.getBAD_REQUEST(
-          "Failed application, account and/or asterisk update.");
+      _logger.info("account deletd " + account.getAccountId());
     }
     
     return Responses.getOK();
